3/5

+++ app/Http/Controllers/Api/AttendeeController.php
@@ -9,9 +9,9 @@ use Illuminate\Http\Request;
 use App\Http\Resources\AttendeeResource;
 class AttendeeController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     */
+    public function __construct(){
+        $this->middleware('auth:sanctum')->except(['index','show']);
+    }
    public function index(Event $event)
 {
     $attendees = $event->attendees()->latest()->paginate();
@@ -45,6 +45,7 @@ class AttendeeController extends Controller
 
     public function destroy( Event $event,Attendee $attendee,)   
     {
+        $this->authorize('delete-attendee', [$event, $attendee]);
         $attendee->delete(); 
         return response(status:204); 
     }
